requirements.txt
aiogram==3.15.0
asyncpg==0.30.0
pyrogram==2.0.106
tgcrypto==1.2.5
yt-dlp==2025.3.31
aiohttp==3.11.14
aiofiles==24.1.0
python-dotenv==1.1.0
psutil==7.0.0
Pillow==11.1.0
mutagen==1.47.0
pydub==0.25.1
